+++ podcasts/ListeningHistoryViewController.swift
@@ -161,18 +161,21 @@ class ListeningHistoryViewController: PCViewController {
     }
 
     @objc func clearTapped() {
-        let optionPicker = OptionsPicker(title: "")
-        let clearAllAction = OptionAction(label: L10n.historyClearAll, icon: nil, action: {
+        let alert = UIAlertController(
+            title: L10n.historyClearAllDetails,
+            message: L10n.historyClearAllDetailsMsg,
+            preferredStyle: .alert
+        )
+        alert.addAction(UIAlertAction(title: L10n.cancel, style: .cancel) { _ in
+            Analytics.track(.listeningHistoryClearConfirmationDismissed)
+        })
+        alert.addAction(UIAlertAction(title: L10n.historyClearAll, style: .destructive) { [weak self] _ in
             Analytics.track(.listeningHistoryCleared)
             DataManager.sharedManager.clearAllEpisodePlayInteractions()
             if SyncManager.isUserLoggedIn() { ServerSettings.setLastClearHistoryDate(Date()) }
-            self.refreshEpisodes(animated: true)
+            self?.refreshEpisodes(animated: true)
         })
-        optionPicker.setNoActionCallback {
-            Analytics.track(.listeningHistoryClearConfirmationDismissed)
-        }
-        optionPicker.addDescriptiveActions(title: L10n.historyClearAllDetails, message: L10n.historyClearAllDetailsMsg, icon: "option-cleanup", actions: [clearAllAction])
-        optionPicker.show(statusBarStyle: preferredStatusBarStyle)
+        present(alert, animated: true)
         Analytics.track(.listeningHistoryClearConfirmationShown)
     }