+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 8.13
 -----
 - Add Liquid Glass support [#4302](https://github.com/Automattic/pocket-casts-ios/pull/4302)
+- Add context menus for Podcasts [#4297](https://github.com/Automattic/pocket-casts-ios/pull/4297)
 - Fix an issue with diagonal swipes not dismissing the player [#4335](https://github.com/Automattic/pocket-casts-ios/pull/4335)
 - Fix player interactive dismiss gesture interaction with vertical scroll views [#4349](https://github.com/Automattic/pocket-casts-ios/pull/4349)